Application ready to simulate in FAST mode. In this mode, optical photons are not propagated. muon detection is done at G4HodoscopeSteppingAction by checking which bars were traversed by muons. Output file returns a hit (1) at corresponding pixels per injected particle.

+++ src/Applications/G4HodoscopeSimulator/G4HodoscopeRunAction.cc
@@ -16,15 +16,23 @@ G4HodoscopeRunAction::G4HodoscopeRunAction(Event& theEvent)
  fEvent(theEvent)
 {
 	G4cout << "...G4HodoscopeRunAction..." << G4endl;
-	string fileName = theEvent.GetSimData().GetOutputFileName();
-	fOutFile = std::ofstream(fileName, ofstream::out | ofstream::app);
-	G4cout << "[DEBUG] G4HodoscopeRunAction: output file name = " << fileName << G4endl;
+
+	if (fEvent.GetSimData().GetSimulationMode() == SimData::SimulationMode::eFull) {
+		cout << "[INFO] G4HodoscopeSimulator::G4HodoscopeRunAction: Running Simulation in Full mode" << endl;
+		string fileName = theEvent.GetSimData().GetOutputFileName();
+		cout << "[INFO] G4HodoscopeSimulator::G4HodoscopeRunAction: opening outputfile: " << fileName << endl;
+		fOutFile = std::ofstream(fileName, ofstream::out | ofstream::app);
+		
+	}
+	
+	
 }
 
 
 G4HodoscopeRunAction::~G4HodoscopeRunAction()
-{
-	fOutFile.close();
+{	
+	if (fEvent.GetSimData().GetSimulationMode() == SimData::SimulationMode::eFull)
+		fOutFile.close();
 }
 
 
@@ -32,23 +40,25 @@ void
 G4HodoscopeRunAction::BeginOfRunAction(const G4Run* aRun)
 {
 
-	// write header only in run = 0.
+	if (fEvent.GetSimData().GetSimulationMode() == SimData::SimulationMode::eFull) {
+		// write header only in run = 0.
 
-	G4int g4RunId = aRun->GetRunID();
-	if (g4RunId == 0) {
-		Detector& currDet = fEvent.GetDetector(0);
-		// loop over OptDevices and print output header
-		// #p_momentum Channel_1 Channel_2 ... Channel_n
-		fOutFile << "# p_momentum ";
-		for (auto odIt = currDet.OptDeviceRange().begin(); odIt != currDet.OptDeviceRange().end(); odIt++) {
-			auto& currOd = odIt->second;
-			int odId = currOd.GetId();
-			//cout << "[DEBUG] Optical Device ID = " << odId << endl;
-			fOutFile << odId << " ";
+		G4int g4RunId = aRun->GetRunID();
+		if (g4RunId == 0) {
+			Detector& currDet = fEvent.GetDetector(0);
+			// loop over OptDevices and print output header
+			// #p_momentum Channel_1 Channel_2 ... Channel_n
+			fOutFile << "# p_momentum ";
+			for (auto odIt = currDet.OptDeviceRange().begin(); odIt != currDet.OptDeviceRange().end(); odIt++) {
+				auto& currOd = odIt->second;
+				int odId = currOd.GetId();
+				//cout << "[DEBUG] Optical Device ID = " << odId << endl;
+				fOutFile << odId << " ";
+			}
+
+			fOutFile << endl;
+		
 		}
-
-		fOutFile << endl;
-	
 	}
 	
 }
@@ -58,6 +68,8 @@ void
 G4HodoscopeRunAction::EndOfRunAction(const G4Run* aRun)
 { 
 
+	if (fEvent.GetSimData().GetSimulationMode() != SimData::SimulationMode::eFull)
+		return;
 	// get current particle information
 	double particleMomentum = G4HodoscopeSimulator::currentParticle.GetMomentum();
 	fOutFile << particleMomentum / CLHEP::MeV << " ";
